typed fixed

+++ prácticas/angular/angular-project/src/app/pages/form/components/form-delete/form-delete.component.ts
@@ -1,17 +1,18 @@
 import { Component, OnInit } from '@angular/core'
 import { FormBuilder, FormGroup, Validators } from '@angular/forms'
 import { ApiService } from 'src/app/services/api.service'
+import { FormDelete } from 'src/app/shared/model/shared'
 
 @Component({
   selector: 'app-form-delete',
   templateUrl: './form-delete.component.html',
   styleUrls: ['./form-delete.component.scss']
 })
-export class FormDeleteComponent implements OnInit {
+export class FormDeleteComponent{
   public userRegisterForm!: FormGroup
-  public posts: any
+  public posts?: Object
   public submitted: boolean = false
-  public deletePosts: any = 0
+  public deletePosts?: FormDelete;
 
   constructor (private request: ApiService, private formBuilder: FormBuilder) {
     this.userRegisterForm = this.formBuilder.group({
@@ -19,24 +20,23 @@ export class FormDeleteComponent implements OnInit {
     })
   }
 
-  ngOnInit (): void {}
-  deletePost = () => {
-    this.request.deleteTodoById(this.deletePosts.id).subscribe(results => {
-      this.posts = results
-      console.log(this.posts)
-    })
+  deletePost ():void {
+    if (this.deletePosts) {
+      this.request.deleteTodoById(this.deletePosts.id).subscribe(results => {
+        this.posts = results
+        console.log(this.posts)
+      })
+    }
   }
 
   public onSubmit (): void {
     this.submitted = true
     if (this.userRegisterForm.valid) {
-      const user: Object = {
+      const user: FormDelete = {
         id: this.userRegisterForm.get('id')?.value
       }
       this.deletePosts = user
       this.deletePost()
-      console.log(user)
-      console.log(this.deletePosts)
       this.userRegisterForm.reset()
       this.submitted = false
     }
